requirements.txt
fastapi>=0.100.0,<0.120.0
uvicorn[standard]>=0.20.0,<0.25.0
google-cloud-bigquery==3.27.0
google-auth==2.37.0
pydantic>=2.7.0,<3.0.0
pydantic-settings>=2.7.0
python-multipart==0.0.20
python-dotenv==1.0.1
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.24.0
pandas==2.2.3
google-oauth2-tool==0.0.3
db-dtypes==1.3.0
pyarrow==17.0.0

# AI/ML Dependencies for LSTM Model
numpy>=1.22.0,<1.24.0
scikit-learn==1.3.0
tensorflow>=2.10.0,<2.14.0
keras>=2.10.0,<2.14.0

# Natural Language Processing
nltk==3.8.1
spacy>=3.5.0,<3.7.0

# Additional utilities
python-dateutil==2.8.2
requests==2.31.0 